Added support for setting lna, mixer and vga gains for the r820t tuner. This is for radio astronomy purposes.  I was finding that the "tuner" gain setting was really ramping up the internal noise of the receiver and I really wanted to get more out of the built in LNA.  I find now I'm able to get more gain by slowly increasing the lna till and then the mixer and then the vga (which doesn't need much before it gets out of control).
What I've done is added arguments for --lna-gain, --mixer-gain and --vga-gain on the command line.  If any one of these is specified and set to 0 or more
then the --gain setting is ignored.  If only the --gain is specified or if no gain options are specified then it all works as before.

This currently relies upon my fork of the librtlsdr library where I added functions to support this for the r820t tuner (can be expended to other tuners easily
enough).  That fork is available here:

https://github.com/mr-growl/librtlsdr

I've submitted a pull request for librtlsdr and if it is accepted I might submit a pull request for this but for now the fork in needed.